refact: init comment refactor
